Moved log cleanup build items to build steps that don't depend on other stuff
